requirements.txt
streamlit
rdkit-pypi
torch
deepchem
pandas
numpy
scikit-learn
transformers
datasets
tqdm
matplotlib
plotly
umap-learn
